fix(ui): dist/styles.css를 직접 임포트 하지 않아도 tailwindCss 적용되도록 수정
