+++ apps/admin/src/app/favicon.ico
