+++ apps/client/.eslintrc.js
@@ -6,4 +6,5 @@ module.exports = {
   parserOptions: {
     project: true,
   },
+  ignorePatterns: ['postcss.config.js', 'tailwind.config.js'],
 };
