+++ apps/client/src/app/favicon.ico
