+++ apps/client/src/app/use/page.tsx
@@ -1,0 +1,5 @@
+const UsesPage = () => {
+  return <div>sdfsdf</div>;
+};
+
+export default UsesPage;
