+++ apps/client/tsconfig.json
@@ -5,7 +5,8 @@
       {
         "name": "next"
       }
-    ]
+    ],
+    "types": ["react", "react-dom"]
   },
   "include": [
     "next-env.d.ts",
