+++ package.json
@@ -28,10 +28,9 @@
     "eslint-plugin-prettier": "^5.0.0",
     "husky": "^9.1.4",
     "lint-staged": "^15.2.8",
+    "nx": "19.5.7",
     "prettier": "^3.3.3",
-    "turbo": "^2.0.12",
-    "typescript": "^5.4.5",
-    "nx": "19.5.7"
+    "turbo": "^2.0.12"
   },
   "packageManager": "pnpm@8.15.6",
   "engines": {
