+++ packages/ui/package.json
@@ -3,13 +3,10 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
-    ".": "./src/index.ts",
-    "./styles": "./dist/tailwind.generated.css"
+    ".": "./src/index.ts"
   },
   "scripts": {
-    "lint": "eslint . --max-warnings 0",
-    "build:css": "postcss src/index.css -o dist/tailwind.generated.css",
-    "build": "pnpm run build:css"
+    "lint": "eslint . --max-warnings 0"
   },
   "devDependencies": {
     "@jjoing/eslint-config": "workspace:*",
@@ -20,12 +17,12 @@
     "@types/react": "^18.2.61",
     "@types/react-dom": "^18.2.19",
     "eslint": "^8.57.0",
+    "postcss": "^8.4.41",
     "postcss-cli": "^11.0.0",
     "typescript": "^5.3.3"
   },
   "dependencies": {
     "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.41",
     "react": "^18.2.0",
     "tailwindcss": "^3.4.9"
   }
