+++ storybook/.storybook/preview.ts
@@ -1,4 +1,3 @@
-import '@jjoing/ui/styles';
 import type { Preview } from '@storybook/react';
 
 const preview: Preview = {
